+++ src/main/java/net/kaikk/mc/fr/ListedItem.java
@@ -4,12 +4,12 @@ import org.bukkit.Material;
 
 class ListedItem {
 	Material material;
-	Byte data;
+	Short data;
 	String world;
 	
 	ListedItem(){}
 	
-	ListedItem(Material material, Byte data, String world) {
+	ListedItem(Material material, Short data, String world) {
 		this.material = material;
 		this.data = data;
 		this.world = world;
@@ -23,7 +23,7 @@ class ListedItem {
 		
 		this.material = (Material) Material.valueOf(arr[0]);
 		if (arr.length>1) {
-			this.data = (arr[1].equals("null")?null:Byte.valueOf(arr[1]));
+			this.data = (arr[1].equals("null")?null:Short.valueOf(arr[1]));
 			if (arr.length>2) {
 				this.world = (arr[2].equals("null")?null:arr[2]);
 			}
@@ -34,12 +34,12 @@ class ListedItem {
 		return material+":"+(data==null&&world==null ? "" : (data==null ? ":*" : ":"+data)+(world==null ? "" : ":"+world));
 	}
 	
-	boolean match(Material material, Byte data, String world) {
+	boolean match(Material material, Short data, String world) {
 		boolean test = material==this.material && (this.data==null || this.data.equals(data)) && (this.world==null || this.world.equals(world));
 		return test;
 	}
 	
-	boolean equals(Material material, Byte data, String world) {
+	boolean equals(Material material, Short data, String world) {
 		boolean test = material==this.material && (this.data==data || (this.data!=null && this.data.equals(data))) && (this.world==world || (this.world!=null && this.world.equals(world)));
 		return test;
 	}
